Add clear screen from cursor down, and from cursor up
Fix clear line left and right of cursor to include the space the cursor
is on.  This mimics what XTerm does.  Also fixes incorrect calculation
where FONTWIDTH was used instead of FONTHEIGHT.

Fix [J and [K to do the same as [0J and [0K

Tested on RC2014 with GFX_USE_DMA set to OFF.  DMA causes a lot of
cursor artifacts.