add 1m,2m,22m and 27m for legacy intensity support
added control codes to control color intensity, for 4 byte color selection in programs that use the legacy color selection scheme 1,2 and 22 (for programs that do not support 90m-107m).  fixed a bug in 7m that would cause older code to occasionally produce output were reverse did not work correctly when calling 7m or 27m when it was not needed (when the state was already swapped). this was done by adding a 'reverse' variable to ctx so that the terminal can keep track of color swaps for the state of the reverse text. the above fixes produce proper output with gorilla.com under CP/M.