Update README.md
 The limited schematics for the Pi Zero do not depict decoupling capacitor on any of the voltage source pins, leaving those sources unregulated for our use. 
I found I needed to add a .1uf capacitor on to the power and ground pins of the PS/2 interface at the Pi GPIO side, and to the 5v and ground pins going to the host computer.  The IBM Model 30-286 uses this method to managing the high current draw the PS/2 keyboard places on the keyboard controller (on the original model 30, it was a different controller and they used chokes, to accomplish this).  the schematics for review can be found at : http://www.walshcomptech.com/ps2schematics/.  IBM also uses a current limiting resistor of 240 ohms, with a modern PS/2 keyboard this might not be necessary my tests without an oscilloscope seam to show it doesn't matter as much as smoothing the current draw.  you would need an oscilloscope to be certain.   Furthermore it may be a good idea to install a fuse on the 3.3v line to prevent a current spike in the event the keyboard is dislodged while the computer is running, though its unlikely it is depicted on IBM's schematics.